worker/requirements.txt
boto3==1.40.67
psycopg2-binary==2.9.11
torch==2.5.1
torchaudio==2.5.1
demucs==4.0.1
yt-dlp
openai-whisper
phonemizer==3.2.1
librosa==0.10.0
soundfile==0.12.1
transformers==4.36.0
numpy==1.24.3
scipy==1.11.4
